requirements.txt
# Core FastAPI Framework
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
aiofiles

# French Language Models & Embeddings
sentence-transformers
transformers
torch
tokenizers

# Vector Database & RAG
faiss-cpu
langchain
langchain-community
numpy

# Document Processing (French Support)
PyMuPDF
python-docx
easyocr
pandas
Pillow

# Logging & Monitoring
python-json-logger

# Optional: For better performance
accelerate

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
structlog==23.2.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# Performance
redis==5.0.1
aiocache==0.12.2
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
locust==2.17.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

openai==1.12.0
python-dotenv==1.0.0